move registration logic to authentication service
