+++ flashcard.wars/src/main/java/study/games/flashcard/wars/service/UserService.java
@@ -1,7 +1,5 @@
 package study.games.flashcard.wars.service;
 
-import org.springframework.security.core.userdetails.UserDetails;
-import study.games.flashcard.wars.models.dtos.UserDto;
 import study.games.flashcard.wars.models.entities.AppUser;
 
 import java.util.List;
@@ -19,8 +17,6 @@ public interface UserService {
     boolean deleteUserById(Long userId);
 
     AppUser updateUser(AppUser appUser);
-
-    AppUser registerUser(UserDto userDto) throws Exception;
 
     List<AppUser> getAllUsers();
